Move dispatcher to application class for mocking in tests
