+++ app/unit-tests/src/org/commcare/utils/MockUtils.kt
@@ -3,13 +3,13 @@ package org.commcare.utils
 import io.mockk.every
 import io.mockk.mockkObject
 import kotlinx.coroutines.Dispatchers
-import org.commcare.tasks.templates.CoroutinesAsyncTask
+import org.commcare.tasks.templates.CoroutineAsyncTaskHelper
 
 /**
  * @author $|-|!˅@M
  */
 fun mockAsyncTaskDispatchers() {
-    mockkObject(CoroutinesAsyncTask)
-    every { CoroutinesAsyncTask.parallelDispatcher() } returns Dispatchers.Main
-    every { CoroutinesAsyncTask.serialDispatcher() } returns Dispatchers.Main
+    mockkObject(CoroutineAsyncTaskHelper)
+    every { CoroutineAsyncTaskHelper.parallelDispatcher() } returns Dispatchers.Main
+    every { CoroutineAsyncTaskHelper.serialDispatcher() } returns Dispatchers.Main
 }
